corregido enviar todos los datos registro

+++ src/main/java/com/glgweb/app/Auth/RegisterRequest.java
@@ -14,7 +14,7 @@ public class RegisterRequest {
     String password;
     String firstname;
     String lastname;
-    Integer phonenumber;
+    Long phonenumber;
     String email;
 
 }
